Action PR: * Rename migration to use slug name in preference over default 'auto' name * Remove schema package and relocate to report package * Fix settings imports to reference django.conf * Use itertools.groupby to extract clinical data summary from cursor for increased clarity * Update demographic fields in Report Designer to be optional * Update report to no longer force patient ID to be exported with results * Update report deletion to use an AJAX Delete request
