--- rdrf/report/admin.py
@@ -1,1 +1,0 @@
-# Register your models here.
