+++ rdrf/report/templates/report_designer.html
@@ -1,5 +1,6 @@
 {% extends "base.html" %}
 {% load i18n %}
+{% load add_attr %}
 {% block content %}
 
 <script type="text/javascript">
@@ -138,7 +139,7 @@
                 {{ form.title.label }} <span class="fa fa-asterisk" style="color: red;" aria-hidden="true"></span>
             </label>
             <div class="col-sm-10">
-                <input type="text" id="title" name="title" class="form-control" value="{{form.title.value|default_if_none:""}}">
+                {{ form.title|add_attr:"class,form-control" }}
             </div>
         </div>
         <div class="row">
@@ -159,7 +160,7 @@
         <div class="row">
             <label for="access_group" class="col-sm-2 col-form-label">{{ form.access_groups.label }}</label>
             <div class="col-sm-10">
-                {{ form.access_groups }}
+                {{ form.access_groups|add_attr:"class,form-select" }}
             </div>
             <div class="offset-sm-2 col-sm-10 form-text">{% trans "The user groups that will have access to this report data." %}</div>
         </div>
@@ -168,7 +169,7 @@
         <legend>Content</legend>
         <div class="row">
             <label for="demographic_fields" class="col-sm-2 col-form-label">
-                {{ form.demographic_fields.label }} <span class="fa fa-asterisk" style="color: red;" aria-hidden="true"></span>
+                {{ form.demographic_fields.label }}
             </label>
             <div class="col-sm-10 mt-2">
                 {{ form.demographic_fields }}
@@ -178,7 +179,7 @@
             <label for="cde_heading_format" class="col-sm-2 col-form-label">
                 {{ form.cde_heading_format.label }}
             </label>
-            <div class="col-sm-10 mt-2">{{ form.cde_heading_format }}</div>
+            <div class="col-sm-10 mt-2">{{ form.cde_heading_format|add_attr:"class,form-select" }}</div>
             <div id="cde_format_label_help" class="offset-sm-2 col-sm-10 form-text">
                 <p>{% trans "CDE Headings are made up of the following parts: Context Form Group, Registry Form, Section, Common Data Element." %}</p>
                 <p>{% trans "This setting determines whether to use the full name label for each of these parts or their respective configured Abbreviated Name." %}</p>
