+++ rdrf/report/templates/reports_list.html
@@ -25,7 +25,19 @@
                 const delete_url = $(e.relatedTarget).data('delete-url');
 
                $("#deleteReportTitle").text(report_title);
-               $("#reportDeleteLink").attr('href', delete_url);
+               $("#btnDeleteReport").data('delete-url', delete_url);
+            });
+
+            $('#btnDeleteReport').on('click', function(e) {
+                const delete_url = $(this).data('delete-url');
+                $.ajax({
+                    type: 'DELETE',
+                    url: delete_url,
+                    headers: {'X-CSRFToken': '{{ csrf_token }}'},
+                    success: function (data) {
+                        window.location.reload();
+                    }
+                });
             });
         });
     </script>
@@ -65,7 +77,7 @@
                     <td>
                         <a href="{% url 'report:report_designer' report.id %}" class="btn btn-sm btn-outline-secondary">{% trans "Edit" %} <i class="fa fa-edit"></i> </a>
                         <button type="button" class="btn btn-sm btn-outline-danger"
-                                data-delete-url="{%  url 'report:report_delete' report.id %}" data-report-title="{{ report.title }}"
+                                data-delete-url="{%  url 'report:report_designer' report.id %}" data-report-title="{{ report.title }}"
                                 data-bs-toggle="modal" data-bs-target="#deleteModal">
                             {% trans "Delete" %} <i class="fa fa-trash-o"></i>
                         </button>
@@ -77,22 +89,24 @@
         </table>
 
         <div id="deleteModal" class="modal" tabindex="-1">
-          <div class="modal-dialog">
-            <div class="modal-content">
-              <div class="modal-header">
-                <h5 class="modal-title">{% trans "Delete Report" %}</h5>
-                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
-              </div>
-              <div class="modal-body">
-                <p>{% trans "Are you sure you want to delete this report?" %}</p>
-                <p>{% trans "Report title:" %} <span id="deleteReportTitle"></span> </p>
-              </div>
-              <div class="modal-footer">
-                <button type="button" class="btn btn-outline-secondary" data-bs-dismiss="modal">Close</button>
-                <a id="reportDeleteLink" href="#" class="btn btn-danger">{% trans "Delete" %} <i class="fa fa-trash-o"></i> </a>
-              </div>
+            <div class="modal-dialog">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title">{% trans "Delete Report" %}</h5>
+                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body">
+                        <p>{% trans "Are you sure you want to delete this report?" %}</p>
+                        <p>{% trans "Report title:" %} <span id="deleteReportTitle"></span> </p>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-outline-secondary" data-bs-dismiss="modal">Close</button>
+                        <button id="btnDeleteReport" type="button" class="btn btn-danger" data-delete-url="">
+                            {% trans "Delete" %} <i class="fa fa-trash-o"></i>
+                        </button>
+                    </div>
+                </div>
             </div>
-          </div>
         </div>
     {% else %}
         <div class="alert alert-info"><p class="mb-0">You don't have any reports to view.</p></div>
