+++ rdrf/report/urls.py
@@ -1,6 +1,6 @@
 from django.urls import re_path
 
-from report.views import ReportDesignView, ReportDeleteView, ReportDownloadView, ReportsView
+from report.views import ReportDesignView, ReportDownloadView, ReportsView
 
 app_name = 'report'
 
@@ -9,7 +9,6 @@ urlpatterns = [
     re_path(r'^list$', ReportsView.as_view(), name='reports_list'),
 
     re_path(r'^(?P<report_id>\w+)/?$', ReportDesignView.as_view(), name='report_designer'),
-    re_path(r'^(?P<report_id>\w+)/delete/?$', ReportDeleteView.as_view(), name='report_delete'),
     re_path(r'^download/(?P<report_id>\w+)/(?P<format>\w+)/?$', ReportDownloadView.as_view(),
             name='report_download'),
